Implement per sample grads util and refactor code
