+++ opacus/tests/grad_samples/dp_multihead_attention_test.py
@@ -20,6 +20,7 @@ from hypothesis import given, settings
 from opacus.layers import DPMultiheadAttention
 
 from .common import GradSampleHooks_test
+from ...utils.per_sample_gradients_utils import check_per_sample_gradients_are_correct, get_grad_sample_modes
 
 
 class DPMultiheadAttentionAdapter(nn.Module):
@@ -53,18 +54,20 @@ class MultiHeadAttention_test(GradSampleHooks_test):
         add_bias_kv=st.booleans(),
         add_zero_attn=st.booleans(),
         kv_dim=st.booleans(),
+        test_or_check=st.integers(1, 2)
     )
     @settings(deadline=10000)
     def test_multihead_attention(
-        self,
-        N: int,
-        T: int,
-        D: int,
-        P: int,
-        bias: bool,
-        add_bias_kv: bool,
-        add_zero_attn: bool,
-        kv_dim: bool,
+            self,
+            N: int,
+            T: int,
+            D: int,
+            P: int,
+            bias: bool,
+            add_bias_kv: bool,
+            add_zero_attn: bool,
+            kv_dim: bool,
+            test_or_check: int
     ):
 
         if kv_dim:
@@ -86,4 +89,9 @@ class MultiHeadAttention_test(GradSampleHooks_test):
         v = torch.randn([T, N, D])
         x = torch.stack((q, k, v), dim=-1)
 
-        self.run_test(x, attn, batch_first=False)
+        if test_or_check == 1:
+            self.run_test(x, attn, batch_first=False)
+        if test_or_check == 2:
+            for grad_sample_mode in get_grad_sample_modes(use_ew=True):
+                assert check_per_sample_gradients_are_correct(x, attn, batch_first=False,
+                                                              grad_sample_mode=grad_sample_mode)
