+++ opacus/tests/grad_samples/linear_test.py
@@ -19,6 +19,7 @@ import torch.nn as nn
 from hypothesis import given, settings
 
 from .common import GradSampleHooks_test
+from ...utils.per_sample_gradients_utils import get_grad_sample_modes, check_per_sample_gradients_are_correct
 
 
 class Linear_test(GradSampleHooks_test):
@@ -30,17 +31,19 @@ class Linear_test(GradSampleHooks_test):
         input_dim=st.integers(2, 4),
         bias=st.booleans(),
         batch_first=st.booleans(),
+        test_or_check=st.integers(1, 2)
     )
     @settings(deadline=10000)
     def test_input_bias(
-        self,
-        N: int,
-        Z: int,
-        W: int,
-        H: int,
-        input_dim: int,
-        bias: bool,
-        batch_first: bool,
+            self,
+            N: int,
+            Z: int,
+            W: int,
+            H: int,
+            input_dim: int,
+            bias: bool,
+            batch_first: bool,
+            test_or_check: int
     ):
 
         if input_dim == 2:
@@ -57,4 +60,10 @@ class Linear_test(GradSampleHooks_test):
         x = torch.randn(x_shape)
         if not batch_first:
             x = x.transpose(0, 1)
-        self.run_test(x, linear, batch_first=batch_first, ew_compatible=N > 0)
+        ew_compatible = N > 0
+        if test_or_check == 1:
+            self.run_test(x, linear, batch_first=batch_first, ew_compatible=ew_compatible)
+        if test_or_check == 2:
+            for grad_sample_mode in get_grad_sample_modes(use_ew=ew_compatible):
+                assert check_per_sample_gradients_are_correct(x, linear, batch_first=batch_first,
+                                                              grad_sample_mode=grad_sample_mode)
